バグ修正 ok

+++ application/controllers/Target.php
@@ -22,7 +22,16 @@ class Target extends CI_Controller {
         $data['content'] = $content;//layout(view)に渡せるように$dataに代入する
         $this->load->view('layout/member/layout', $data);//layout/layout.phpを表示させる
     }
-
+    
+    /**
+    * 引数に整数のみ受け付ける条件
+    */
+    private function argumentCheck($year, $term)//if文の条件を共通化
+    {
+        return !is_numeric($year, $term) || intval($year, $term) < 1;//returnしないと正常に動かない。
+    }
+    
+    
     /**
      * 目標一覧
      */
@@ -39,9 +48,14 @@ class Target extends CI_Controller {
      */
     public function add() 
     {
-        $this->form_validation->set_rules('year', '年度', 'required|regex_match[/^[0-9]{4}$/]');
-        $this->form_validation->set_rules('term', '期間', 'required');
+        $this->form_validation->set_rules('year', '年度', 'required|regex_match[/^[0-9]{4}$/]|callback_year_add_check');
+        $this->form_validation->set_rules('term', '期間', 'required|callback_term_add_check');
         $this->form_validation->set_rules('target', '目標', 'required');
+        //validationメッセージ
+        $this->form_validation->set_message('required', '【{field}】が未入力です');
+        $this->form_validation->set_message('regex_match', '【{field}】の入力形式が違います');
+        $this->form_validation->set_message('year_add_check', 'この【{field}】の目標は四半期全て入力済みです');
+        $this->form_validation->set_message('term_add_check', 'この【{field}】の目標は既に入力済みです');
         
         if ($this->form_validation->run() === FALSE) {//target/addにアクセスした際、formの入力がないためviewが表示できる。
             $this->showView('target/add');//viewを返す
@@ -53,6 +67,7 @@ class Target extends CI_Controller {
             ];
             $member_id = $this->session->userdata('member_id');//sessionのデータを変数$member_idに格納する。
             $this->Target_model->create($target, $member_id);//dbへtargetの内容とmember_idを登録する。
+            $this->session->set_flashdata('flash_message', '新しい目標を追加しました');
             redirect("target/index");//各ユーザのindexページへリダイレクトする。
         }
     }
@@ -60,31 +75,46 @@ class Target extends CI_Controller {
     /**
      *  目標の編集
      */
-    public function edit($year, $term)//とりあえず$member_idに初期値0を設定。
+    public function edit($year, $term)
     {
-        $this->form_validation->set_rules('year', '年度', 'required|regex_match[/^[0-9]{4}$/]');
-        $this->form_validation->set_rules('term', '期間', 'required');
+        $this->form_validation->set_rules('year', '年度', 'required|regex_match[/^[0-9]{4}$/]|callback_year_edit_check');
+        $this->form_validation->set_rules('term', '期間', 'required|callback_term_edit_check');
         $this->form_validation->set_rules('target', '目標', 'required');
         
-        if ($this->form_validation->run() === FALSE) {
-            $member_id = $this->session->userdata('member_id');
-            $data['target'] = $this->Target_model->findByTarget($member_id, $year, $term);//編集するtargetの既存データを取得する。
-            $this->showView('target/edit', $data);//$data['target']を引数にviewを返す。
+        //validationメッセージ
+        $this->form_validation->set_message('required', '【{field}】が未入力です');
+        $this->form_validation->set_message('regex_match', '【{field}】の入力形式が違います');
+        $this->form_validation->set_message('year_edit_check', 'この【{field}】の目標は四半期全て入力済みです');
+        $this->form_validation->set_message('term_edit_check', 'この【{field}】の目標は既に入力済みです');
+        
+        if ($this->argumentCheck($year, $term)) {
+            redirect('member/logout');
         } else {
-            $member_id = $this->session->userdata('member_id');
-            $findTarget = [//target特定用データ
-                'member_id' => $member_id,
-                'year' => $year,
-                'term' => $term
-            ];
-            $updateTarget = [//上書きするデータを取得する。
-                'year' => $this->input->post('year'),//inputで取得したyearデータ
-                'term' => $this->input->post('term'),//inputで取得したtermデータ
-                'target' => $this->input->post('target')//inputで取得したtargetデータ
-            ];
-            $this->Target_model->update($updateTarget, $findTarget);//dbのデータを引数で上書きする。
-            redirect('target/index');//リダイレクトさせる。
+            if ($this->form_validation->run() === FALSE) {
+                $member_id = $this->session->userdata('member_id');
+                $this->session->set_userdata('year', $year);
+                $this->session->set_userdata('term', $term);
+                var_dump($this->session->userdata('year'));
+                $data['target'] = $this->Target_model->findByTarget($member_id, $year, $term);//編集するtargetの既存データを取得する。
+                $this->showView('target/edit', $data);//$data['target']を引数にviewを返す。
+            } else {
+                $member_id = $this->session->userdata('member_id');
+                $findTarget = [//target特定用データ
+                    'member_id' => $member_id,
+                    'year' => $year,
+                    'term' => $term
+                ];
+                $updateTarget = [//上書きするデータを取得する。
+                    'year' => $this->input->post('year'),//inputで取得したyearデータ
+                    'term' => $this->input->post('term'),//inputで取得したtermデータ
+                    'target' => $this->input->post('target')//inputで取得したtargetデータ
+                ];
+                $this->Target_model->update($updateTarget, $findTarget);//dbのデータを引数で上書きする。
+                $this->session->set_flashdata('flash_message', '目標を編集しました');
+                redirect('target/index');//リダイレクトさせる。
+            }
         }
+        
     }
     
     /**
@@ -92,8 +122,80 @@ class Target extends CI_Controller {
      */
     public function delete($year, $term) 
     {
+        if ($this->argumentCheck($year, $term)) {
+            redirect('member/logout');
+        } else {
+            $member_id = $this->session->userdata('member_id');
+            $this->Target_model->destroy($member_id, $year, $term);
+            $this->session->set_flashdata('flash_message', '目標を削除しました');//削除時のflash
+            redirect('target/index');
+        }
+    }
+    
+    /**
+     * コールバック処理
+     * 目標登録時のyear_validation
+     */
+    public function year_add_check($year)
+    {
+        $member_id = $this->session->userdata('member_id');//member_idを取得する
+        $targets = $this->Target_model->findByIdAndYear($member_id, $year);//post時のtargetを連想配列で受け取る
+        if (count($targets) < 4) {//db内にtargetが3つ以下であれば通す
+            return TRUE;
+        } else {
+            return FALSE;
+        }   
+    }
+    
+    /**
+     * 目標編集時のyear_validation
+     */
+    public function year_edit_check($year)
+    {
         $member_id = $this->session->userdata('member_id');
-        $this->Target_model->destroy($member_id, $year, $term);
-        redirect("target/index");
+        $targets = $this->Target_model->findByIdAndYear($member_id, $year);//post時のtargetを連想配列で受け取る
+        $postBeforeYear = $this->session->userdata('year');//post前のyearをeditの引数からsessionを通して取得する
+        foreach ($targets as $target) {
+            //返り値の配列が4以下でpost前とpost時とyearが同じ場合(何も変更しないで編集した場合) || 返り値の配列が3以下でpost前とpost時のyearが違う場合
+           if ((count($targets) < 5 && $year === $postBeforeYear ) || (count($targets) < 4 && $year !== $postBeforeYear)) {
+                return TRUE;
+            } else {
+                return FALSE;
+            }   
+        }
+    }
+    
+    /**
+     * 目標登録時のterm_validation
+     */
+    public function term_add_check($term)
+    {
+        $member_id = $this->session->userdata('member_id');
+        $postYear = $this->input->post('year');//post時のyearを取得する
+        $target = $this->Target_model->findByTarget($member_id, $postYear, $term);//post時のmember_idとyearとtermで取得する
+        if (!$targec) {//存在しなければ通す
+            return TRUE;
+        } else {
+            return FALSE;
+        }
+    }
+    
+    /**
+     * 目標編集時のterm_validation
+     */
+    public function term_edit_check($term)
+    {
+        $member_id = $this->session->userdata('member_id');
+        $postYear = $this->input->post('year');
+        $target = $this->Target_model->findByTarget($member_id, $postYear, $term);//post時のtargetを取得する
+        $postBeforeYear = $this->session->userdata('year');//post前のyear
+        $postBeforeTerm = $this->session->userdata('term');//post前のterm
+        var_dump($postBeforeTerm);
+        //編集せずにpostした場合　|| post時のmember_idとpostYearとtermで検索しtargetがない場合
+        if ($term === $postBeforeTerm && $postYear === $postBeforeYear || !$target) {
+            return TRUE;
+        } else {
+            return FALSE;
+        }
     }
 }
